Client - Reformatted Codebase.

+++ Client/src/NodeList.java
@@ -4,113 +4,90 @@
 
 final class NodeList {
 
-    public NodeList()
-    {
+    private final Node head;
+    private Node current;
+
+    public NodeList() {
         head = new Node();
         head.prev = head;
         head.next = head;
     }
 
-    public void insertHead(Node node)
-    {
-        if(node.next != null)
-            node.unlink();
+    public void insertHead(Node node) {
+        if (node.next != null) node.unlink();
         node.next = head.next;
         node.prev = head;
         node.next.prev = node;
         node.prev.next = node;
     }
 
-    public void insertTail(Node node)
-    {
-        if(node.next != null)
-            node.unlink();
+    public void insertTail(Node node) {
+        if (node.next != null) node.unlink();
         node.next = head;
         node.prev = head.prev;
         node.next.prev = node;
         node.prev.next = node;
     }
 
-    public Node popHead()
-    {
+    public Node popHead() {
         Node node = head.prev;
-        if(node == head)
-        {
+        if (node == head) {
             return null;
-        } else
-        {
+        } else {
             node.unlink();
             return node;
         }
     }
 
-    public Node reverseGetFirst()
-    {
+    public Node reverseGetFirst() {
         Node node = head.prev;
-        if(node == head)
-        {
+        if (node == head) {
             current = null;
             return null;
-        } else
-        {
+        } else {
             current = node.prev;
             return node;
         }
     }
 
-    public Node getFirst()
-    {
+    public Node getFirst() {
         Node node = head.next;
-        if(node == head)
-        {
+        if (node == head) {
             current = null;
             return null;
-        } else
-        {
+        } else {
             current = node.next;
             return node;
         }
     }
 
-    public Node reverseGetNext()
-    {
+    public Node reverseGetNext() {
         Node node = current;
-        if(node == head)
-        {
+        if (node == head) {
             current = null;
             return null;
-        } else
-        {
+        } else {
             current = node.prev;
             return node;
         }
     }
 
-    public Node getNext()
-    {
+    public Node getNext() {
         Node node = current;
-        if(node == head)
-        {
+        if (node == head) {
             current = null;
             return null;
         }
         current = node.next;
-            return node;
+        return node;
     }
 
-    public void removeAll()
-    {
-        if(head.prev == head)
-            return;
-        do
-        {
+    public void removeAll() {
+        if (head.prev == head) return;
+        do {
             Node node = head.prev;
-            if(node == head)
-                return;
+            if (node == head) return;
             node.unlink();
-        } while(true);
+        } while (true);
     }
-
-    private final Node head;
-    private Node current;
 }
